Reduce Density & Biomass
In recent runs, was getting an error from having drained the water
table.  I think this is from putting more biomass on the landscape than
the landscape could handle.  Solution: reduce the number of plants &
biomass proportionate to the number of time steps you’re using to
populate the initial conditions